pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data_classifier"
version = "0.1.0"
description = "General-purpose data classification engine for detecting sensitive data in structured database columns."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "httpx>=0.27",
    "agile-backlog>=0.30",
]

[tool.setuptools.packages.find]
include = ["data_classifier*"]

[tool.setuptools.package-data]
data_classifier = ["profiles/*.yaml"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["N815"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
